# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: GovBid Pro API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/webhooks/github:
    post:
      summary: Receive a GitHub webhook delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSummaryResponse"
...
  /api/webhooks/github/events/{id}:
    get:
      summary: Get a stored GitHub webhook event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEventResponse"
        "404":
          description: Not found
components:
  schemas:
...
    WebhookSummary:
      type: object
      properties:
        message:
          type: string
        repo:
          type: string
        sender:
          type: string
        event:
          type: string
        action:
          type: string
          nullable: true
        metadata:
          type: object
    WebhookSummaryResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/WebhookSummary"
    WebhookEvent:
      type: object
      properties:
        id:
          type: string
        event:
          type: string
        action:
          type: string
          nullable: true
        repository:
          type: string
          nullable: true
        sender:
          type: string
          nullable: true
        deliveredAt:
          type: string
        summary:
          $ref: "#/components/schemas/WebhookSummary"
    WebhookEventResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/WebhookEvent"
...